chore(vdc): Add schema/types
