+++ internal/provider/vdc/acl_types.go
@@ -1,0 +1,10 @@
+package vdc
+
+import "github.com/hashicorp/terraform-plugin-framework/types"
+
+type aclResourceModel struct {
+	ID                  types.String `tfsdk:"id"`
+	VDC                 types.String `tfsdk:"vdc"`
+	EveryoneAccessLevel types.String `tfsdk:"everyone_access_level"`
+	SharedWith          types.Set    `tfsdk:"shared_with"`
+}
